+++ Example/Tests/SpireTests.swift
@@ -35,7 +35,7 @@ class SpireTestsSwift: BaseTestsSwift, WDManagerDelegate
             NSLog("\n\t\t    [%@ %@] Not runnable on simulator 📱",file.deletingPathExtension , #function);
             return
             
-        #else
+        #endif
         
         //PART 1: We log-in and request user data
         //--------------------------------------
@@ -75,10 +75,24 @@ class SpireTestsSwift: BaseTestsSwift, WDManagerDelegate
             XCTFail("Error when updating the terminal. Make sure your terminal is paired in your iOS device settings and that the terminal is in stand-by mode (ie. by switching off and then on and waiting until the screen lights off).")
         }
         
+        let doPurchaseAndRefund = true //Change this to false to do card authorised and capture of payments instead of regular purchase + refund
+        
+        if doPurchaseAndRefund
+        {
+            self.doPurchaseAndRefund()
+        }
+        else
+        {
+            self.doAuthorisedAndCapture()
+        }
+    }
+    
+    func doPurchaseAndRefund()
+    {
         //PART 4: We do a card sale using Spire terminal
         //---------------------------------------------
         expectation = self.expectation(description: "Card sale")
-        self.doCardPayment()
+        self.doCardPayment(authorised: false)
         self.waitForExpectations(timeout: 300, handler: nil)
         if (self.saleResponse == nil)
         {
@@ -89,7 +103,7 @@ class SpireTestsSwift: BaseTestsSwift, WDManagerDelegate
         {
             SaleHelper.sharedInstance().saleToSaveId(from:self.saleResponse)
         }
-
+        
         //PART 5: We refund the sale we just did
         //(for testing purposes; this is important if you develop using an actual credit card)
         //Algo note that refunds are allowed only for some period of time (ie: midnight on current time, or next 24 hours, etc) and after that, only reversals are allowed. But refunds are for free, while reversals actually cost money (so use reversals responsibly!)
@@ -97,7 +111,7 @@ class SpireTestsSwift: BaseTestsSwift, WDManagerDelegate
         expectation = self.expectation(description: "Refund sale")
         self.refundTransaction()
         self.waitForExpectations(timeout: 300, handler: nil)
-        if let response = self.saleResponse, response.status != .returned
+        if let response = self.saleResponse, response.status != .returned && response.status != .partiallyReturned
         {
             XCTFail("Sale did not succeed. Make sure your terminal is paired in your iOS device settings and that the terminal is in stand-by mode (ie. by switching off and then on and waiting until the screen lights off).")
         }
@@ -105,9 +119,41 @@ class SpireTestsSwift: BaseTestsSwift, WDManagerDelegate
         {
             XCTFail("Sale response empty. Make sure your terminal is paired in your iOS device settings and that the terminal is in stand-by mode (ie. by switching off and then on and waiting until the screen lights off).")
         }
-        #endif
     }
     
+    func doAuthorisedAndCapture()
+    {
+        //PART 4: We do a card sale using Spire terminal
+        //---------------------------------------------
+        expectation = self.expectation(description: "Card sale")
+        self.doCardPayment(authorised: true)
+        self.waitForExpectations(timeout: 300, handler: nil)
+        if (self.saleResponse == nil)
+        {
+            XCTFail("Sale did not succeed. Make sure your terminal is paired in your iOS device settings and that the terminal is in stand-by mode (ie. by switching off and then on and waiting until the screen lights off).")
+            //NOTE:  if your merchant settings have cash mgmt enabled in backend, you will need to run cash tests first - otherwise you will receive a "not authorized" kind of error
+        }
+        else
+        {
+            SaleHelper.sharedInstance().saleToSaveId(from:self.saleResponse)
+        }
+        
+        //PART 5: We capture the sale we just did
+        //(for testing purposes; this is important if you develop using an actual credit card)
+        //Algo note that refunds are allowed only for some period of time (ie: midnight on current time, or next 24 hours, etc) and after that, only reversals are allowed. But refunds are for free, while reversals actually cost money (so use reversals responsibly!)
+        //-------------------------------------------
+        expectation = self.expectation(description: "Refund sale")
+        self.captureSale()
+        self.waitForExpectations(timeout: 300, handler: nil)
+        if let response = self.saleResponse, let payment = response.payments.first as? WDPaymentDetail, payment.status != .captured
+        {
+            XCTFail("Sale did not succeed. Make sure your terminal is paired in your iOS device settings and that the terminal is in stand-by mode (ie. by switching off and then on and waiting until the screen lights off).")
+        }
+        else if self.saleResponse == nil
+        {
+            XCTFail("Sale response empty. Make sure your terminal is paired in your iOS device settings and that the terminal is in stand-by mode (ie. by switching off and then on and waiting until the screen lights off).")
+        }
+    }
     
     func checkingForTerminalConfigUpdates()
     {
@@ -135,9 +181,8 @@ class SpireTestsSwift: BaseTestsSwift, WDManagerDelegate
                                    completion:completionUpdate)
     }
     
-    func doCardPayment()
+    func doCardPayment(authorised : Bool)
     {
-        
         guard let sale = SaleHelper.sharedInstance().newSale() else
         {
             XCTFail("Something went really wrong - doCardPayment")
@@ -145,18 +190,19 @@ class SpireTestsSwift: BaseTestsSwift, WDManagerDelegate
             return
         }
         self.aSale = sale
+        self.aSale.cashRegisterId = UserHelper.sharedInstance()?.selectedCashRegisterId() ?? ""
         self.aSale.addSaleItem(NSDecimalNumber(value: 3.4),
-                               quantity:NSDecimalNumber(value: 5),
+                               quantity:NSDecimalNumber.init(value: 5),
                                taxRate:UserHelper.sharedInstance().preferredSaleItemTax(),
                                itemDescription:"Red Apple",
                                productId:"Dummy ID 1",
-                               externalProductId:nil)
+                               externalProductId : nil)
         self.aSale.addSaleItem(NSDecimalNumber(value: 2.25),
-                               quantity:NSDecimalNumber(value: 3),
+                               quantity:NSDecimalNumber.init(value: 3),
                                taxRate:UserHelper.sharedInstance().preferredSaleItemTax(),
                                itemDescription:"Orange",
                                productId:"Dummy ID 2",
-                               externalProductId:nil)
+                               externalProductId : nil)
         //You can add a service charge to the whole basket -- but this is optional
         self.aSale.addServiceCharge(UserHelper.sharedInstance().serviceChargeRate(),
                                     taxRate:UserHelper.sharedInstance().serviceChargeTax())
@@ -165,11 +211,20 @@ class SpireTestsSwift: BaseTestsSwift, WDManagerDelegate
                                taxRate:UserHelper.sharedInstance().tipTax())
         //You can add a discount for the whole basket when productId is nil, or per productId otherwise. Below, a discount of 6%
         self.aSale.addFlatDiscount(NSDecimalNumber(value: 6.0))
-        self.aSale.cashRegisterId = UserHelper.sharedInstance().selectedCashRegisterId() //Note: if your backend settings have cash mgmt enabled in backend, you will need to run cash tests first to get this value as well as shiftId below
+    //    self.aSale.cashRegisterId = UserHelper.sharedInstance().selectedCashRegisterId() //Note: if your backend settings have cash mgmt enabled in backend, you will need to run cash tests first to get this value as well as shiftId below
         self.aSale.shiftId = UserHelper.sharedInstance().lastShiftId()
         self.aSale.resetPayments()
-        self.aSale.addCardPayment(self.aSale.totalToPay() ?? NSDecimalNumber.init(value:0),
-                                                 terminal:self.selectedDevice!)
+        
+        if authorised
+        {
+            self.aSale.addCardAuthorization(self.aSale.totalToPay() ?? NSDecimalNumber.init(value:0),
+                                      terminal:self.selectedDevice!)
+        }
+        else
+        {
+            self.aSale.addCardPayment(self.aSale.totalToPay() ?? NSDecimalNumber.init(value:0),
+                                      terminal:self.selectedDevice!)
+        }
         
         if let paymentConfiguration : WDSaleRequestConfiguration = WDSaleRequestConfiguration.init(saleRequest: self.aSale)
         {
@@ -187,8 +242,9 @@ class SpireTestsSwift: BaseTestsSwift, WDManagerDelegate
             self.expectation.fulfill()
             return
         }
-        //This is an example of full refund: all items are refunded, EXCEPT SERVICE CHARGE
+        //This is an example of full refund: all items are refunded, EXCEPT SERVICE CHARGE and GRATUITY
         saleToBeRefunded.removeServiceCharge()
+        saleToBeRefunded.removeGratuity()
         guard let totalToPay = saleToBeRefunded.totalToPay() else
         {
             XCTFail("Something went really wrong - refundTransaction totalToPay")
@@ -196,6 +252,7 @@ class SpireTestsSwift: BaseTestsSwift, WDManagerDelegate
             return
         }
         saleToBeRefunded.cashRegisterId = UserHelper.sharedInstance().selectedCashRegisterId()
+        //saleToBeRefunded.cashierId = self.aSale.cashierId
         saleToBeRefunded.customerId = self.aSale.customerId
         //This is an example of full refund: all items are refunded, EXCEPT SERVICE CHARGE
         saleToBeRefunded.removeServiceCharge()
@@ -208,7 +265,30 @@ class SpireTestsSwift: BaseTestsSwift, WDManagerDelegate
             self?.expectation.fulfill()
         })
     }
+    
+    func captureSale()
+    {
+        guard let saleToBeCaptured = self.saleResponse else
+        {
+            XCTFail("Something went really wrong - no sale to capture")
+            self.expectation.fulfill()
+            return
+        }
+        
+        
+        let refSaleRequest = saleToBeCaptured.referenceSaleRequest()
+        
+        refSaleRequest.addCardCapture(saleToBeCaptured.totalToPay() ?? NSDecimalNumber.one, originalPaymentId: saleToBeCaptured.processedCardPayments().first?.internalId ?? "")
+        
+        if let paymentConfiguration = WDSaleRequestConfiguration(saleRequest: refSaleRequest)
+        {
+            sdk.terminalManager.setActive(self.selectedDevice, completion:{[weak self]() in
+                self?.sdk.saleManager.pay(paymentConfiguration, with: (self?.paymentHandler)!)
+            })
+        }
+    }
 
+    
     
     func device(_ device: WDTerminal, connectionStatusDidChange status:WDExtensionConnectionStatus)
     {